% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_data_dictionary.R
\name{align_data_dictionary}
\alias{align_data_dictionary}
\title{Align the data dictionary with the data}
\usage{
align_data_dictionary(
  path_to_cur_dat,
  path_to_dat_dict,
  undefined_vars = NULL,
  required_override = NULL,
  dttm_cols = NULL,
  date_cols = NULL,
  num_cols = NULL
)
}
\arguments{
\item{path_to_cur_dat}{Path to the data we're aligning to.}

\item{path_to_dat_dict}{Path to the data dictionary to be aligned.}

\item{undefined_vars}{Passed on to \code{add_undefined_vars()}, default of NULL uses a built-in list.}

\item{required_override}{Variables to manually set to be required.  Defaults to  'record_id', 'redcap_repeat_instrument' and 'redcap_repeat_instance'.}

\item{dttm_cols}{Columns to mark as datetime, if NULL uses default in \code{dd_assign_coltypes()}.}

\item{date_cols}{Columns to mark as date, if NULL uses default in \code{dd_assign_coltypes()}.}

\item{num_cols}{Columns to mark as numeric if NULL uses default in \code{dd_assign_coltypes()}.}
}
\value{
Aligned data dictionary (a tibble).
}
\description{
Align the data dictionary with the data
}
