% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dv_tab_ca_dx.R
\name{build_dv_tab_ca_dx}
\alias{build_dv_tab_ca_dx}
\title{Build derived variable table for cancer diagnosis form}
\usage{
build_dv_tab_ca_dx(
  tab,
  dat_dict_sub,
  cast_to_double = c("tr_eligible", "age_dx")
)
}
\arguments{
\item{tab}{Cancer diagnosis form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for this instrument.}

\item{cast_to_double}{Character vector of column names to coerce to
\code{double} at the end of the function. Defaults to \code{"tr_eligible"}.}
}
\value{
A dataframe similar to ca_dx with additional derivations.
}
\description{
Build derived variable table for cancer diagnosis form
}
\examples{
# Going to need synthetic data probably...
}
