% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dv_tab_img.R
\name{build_dv_tab_img}
\alias{build_dv_tab_img}
\title{Build derived variable table for imaging form}
\usage{
build_dv_tab_img(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Imaging form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for
this instrument.}
}
\value{
A dataframe similar to img with additional derivations.
}
\description{
Build derived variable table for imaging form
}
\examples{
# build_dv_tab_img(img, dat_dict_sub)
}
