% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dv_tab_reg.R
\name{build_dv_tab_reg}
\alias{build_dv_tab_reg}
\title{Build derived variable table for cancer-directed regimen form}
\usage{
build_dv_tab_reg(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Cancer-directed drugs (regimen) form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for this instrument.}
}
\value{
A dataframe similar to regimen with additional derivations.
}
\description{
Build derived variable table for cancer-directed regimen form
}
\examples{
# Going to need synthetic data probably...
}
