% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_exclusion_helper_derived.R
\name{column_exclusion_helper_derived}
\alias{column_exclusion_helper_derived}
\title{Column exclusion helper for derived variables.}
\usage{
column_exclusion_helper_derived(dat)
}
\arguments{
\item{dat}{Any dataset with PRISSMM variable names (otherwise it won't do much.}
}
\value{
A version of \code{dat} with column names removed that we typically don't share.
}
\description{
Column exclusion helper for derived variables.
}
