% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mos_yrs_intervals.R
\name{days_to_mos}
\alias{days_to_mos}
\title{Convert days to months}
\usage{
days_to_mos(x, days_per_mo = 30.4)
}
\arguments{
\item{x}{Numeric vector of days.}

\item{days_per_mo}{Conversion constant. Default 30.4 (from OLD_CODE).}
}
\value{
Numeric vector of months.
}
\description{
Convert days to months
}
