% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_readr.R
\name{dd_readr}
\alias{dd_readr}
\title{Read and clean up a PRISSMM REDCap data dictionary}
\usage{
dd_readr(dat_dict_path)
}
\arguments{
\item{dat_dict_path}{path to the data dictionary csv}
}
\value{
a data dictionary with renamed columns and other small improvements
}
\description{
Read and clean up a PRISSMM REDCap data dictionary
}
