% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_cpt.R
\name{derive_cpt}
\alias{derive_cpt}
\title{Derive CPT (cancer panel test) form}
\usage{
derive_cpt(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Cancer panel test form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for
this instrument.}
}
\value{
A dataframe similar to cpt with additional derivations.
}
\description{
Derive CPT (cancer panel test) form
}
\examples{
# derive_cpt(cpt, dat_dict_sub)
}
