% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_img.R
\name{derive_img}
\alias{derive_img}
\title{Derive imaging form}
\usage{
derive_img(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Imaging form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for
this instrument.}
}
\value{
A dataframe similar to img with additional derivations.
}
\description{
Derive imaging form
}
\examples{
# derive_img(img, dat_dict_sub)
}
