% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_med_onc.R
\name{derive_med_onc}
\alias{derive_med_onc}
\title{Derive medical oncologist form}
\usage{
derive_med_onc(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Medical oncologist form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for
this instrument.}
}
\value{
A dataframe similar to med_onc with additional derivations.
}
\description{
Derive medical oncologist form
}
\examples{
# derive_med_onc(med_onc, dat_dict_sub)
}
