% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_path.R
\name{derive_path}
\alias{derive_path}
\title{Derive pathology form}
\usage{
derive_path(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Pathology form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for
this instrument.}
}
\value{
A dataframe similar to path with additional derivations.
}
\description{
Derive pathology form
}
\examples{
# derive_path(path, dat_dict_sub)
}
