% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_path_insitu_any.R
\name{derive_path_insitu_any}
\alias{derive_path_insitu_any}
\title{Derive path_insitu_any (any specimen with in situ disease)}
\usage{
derive_path_insitu_any(dat)
}
\arguments{
\item{dat}{A data frame containing mapped \verb{path_insitu*} columns.}
}
\value{
\code{dat} with \code{path_insitu_any} added.
}
\description{
Returns \code{"Yes"} if any \verb{path_insitu*} column is \code{"Yes"}, otherwise \code{"No"}.
Should be called after \code{\link[=map_path_insitu]{map_path_insitu()}} has been applied.
}
\examples{
# derive_path_insitu_any(path)
}
