% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_pt.R
\name{derive_pt}
\alias{derive_pt}
\title{Derive the patient data table.}
\usage{
derive_pt(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Curated pt table.}

\item{dat_dict_sub}{Data dictionary with columns relevant to pt table.}
}
\value{
Derived patient table.
}
\description{
Derive the patient data table.
}
