% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_rad.R
\name{derive_rad}
\alias{derive_rad}
\title{Derive radiation therapy form}
\usage{
derive_rad(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Radiation therapy form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for
this instrument.}
}
\value{
A dataframe similar to rad with additional derivations.
}
\description{
Derive radiation therapy form
}
\examples{
# derive_rad(rad, dat_dict_sub)
}
