% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_reg.R
\name{derive_reg}
\alias{derive_reg}
\title{Derive cancer-directed regimen form}
\usage{
derive_reg(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Cancer-directed drugs (regimen) form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for this instrument.}
}
\value{
A dataframe similar to regimen with additional derivations.
}
\description{
Derive cancer-directed regimen form
}
\examples{
# Going to need synthetic data probably...
}
