% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_tum.R
\name{derive_tum}
\alias{derive_tum}
\title{Derive tumor marker form}
\usage{
derive_tum(tab, dat_dict_sub)
}
\arguments{
\item{tab}{Tumor marker form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for
this instrument.}
}
\value{
A dataframe similar to tum with additional derivations.
}
\description{
Derive tumor marker form
}
\examples{
# derive_tum(tum, dat_dict_sub)
}
