% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stubs_in_data.R
\name{find_stubs_in_data}
\alias{find_stubs_in_data}
\title{Find variables with sub names in a list of columns}
\usage{
find_stubs_in_data(dat_cols)
}
\arguments{
\item{dat_cols}{A vector of column names}
}
\value{
A tibble with columns stub (the root), min and max (the highest and lowest appended number after the stub).
}
\description{
Find variables with sub names in a list of columns
}
