% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_clin_t1_det.R
\name{map_ca_clin_t1_det}
\alias{map_ca_clin_t1_det}
\title{Map ca_clin_t1_det codes to text labels}
\usage{
map_ca_clin_t1_det(x)
}
\arguments{
\item{x}{A numeric (or character) vector of ca_clin_t1_det codes.}
}
\value{
A character vector of clinical T1 detail labels.
}
\description{
Map ca_clin_t1_det codes to text labels
}
\examples{
map_ca_clin_t1_det(c(7, 9, 15, 99))
}
