% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_clin_t_stage.R
\name{map_ca_clin_t_stage}
\alias{map_ca_clin_t_stage}
\title{Map ca_clin_t_stage codes to text labels}
\usage{
map_ca_clin_t_stage(x)
}
\arguments{
\item{x}{A numeric (or character) vector of ca_clin_t_stage codes.}
}
\value{
A character vector of clinical T stage labels.
}
\description{
Map ca_clin_t_stage codes to text labels
}
\examples{
map_ca_clin_t_stage(c(1, 8, 40, 99))
}
