% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_stage.R
\name{map_ca_stage}
\alias{map_ca_stage}
\title{Map ca_stage codes to text labels}
\usage{
map_ca_stage(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{ca_stage} codes.}
}
\value{
A character vector of labels.
}
\description{
Map ca_stage codes to text labels
}
\examples{
map_ca_stage(c(0, 1, 13, 98, 99))
}
