% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ca_stage_iv.R
\name{map_ca_stage_iv}
\alias{map_ca_stage_iv}
\title{Map ca_stage_iv codes to text labels}
\usage{
map_ca_stage_iv(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{ca_stage_iv} codes.}
}
\value{
A character vector of labels.
}
\description{
Map ca_stage_iv codes to text labels
}
\examples{
map_ca_stage_iv(c(0, 1, 2, 9))
}
