% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_image_ca.R
\name{map_image_ca}
\alias{map_image_ca}
\title{Map image_ca codes to text labels}
\usage{
map_image_ca(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{image_ca} codes.}
}
\value{
A character vector of labels.
}
\description{
Map image_ca codes to text labels
}
\examples{
map_image_ca(c(1, 2, 3, 4))
}
