% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_image_casite.R
\name{map_image_casite}
\alias{map_image_casite}
\title{Map image_casite codes to ICD-O-3 site labels}
\usage{
map_image_casite(x, whitespace_match_legacy = TRUE)
}
\arguments{
\item{x}{A numeric (or character) vector of site codes.}

\item{whitespace_match_legacy}{If \code{TRUE}, use four spaces between the ICD
code and description to match the legacy data format. Default \code{FALSE}.}
}
\value{
A character vector of labels.
}
\description{
Uses the metastasis site mapping from \code{\link[=load_met_site_map]{load_met_site_map()}} to recode
\code{image_casite} variables.
}
\examples{
map_image_casite(c(349, 809))
}
