% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_image_scan_type.R
\name{map_image_scan_type}
\alias{map_image_scan_type}
\title{Map image_scan_type codes to text labels}
\usage{
map_image_scan_type(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{image_scan_type} codes.}
}
\value{
A character vector of labels.
}
\description{
Map image_scan_type codes to text labels
}
\examples{
map_image_scan_type(c(1, 3, 5))
}
