% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_naaccr_sex_code.R
\name{map_naaccr_sex_code}
\alias{map_naaccr_sex_code}
\title{Map NAACCR sex codes to text labels}
\usage{
map_naaccr_sex_code(x)
}
\arguments{
\item{x}{A numeric (or character) vector of NAACCR sex codes.}
}
\value{
A character vector of sex labels.
}
\description{
Map NAACCR sex codes to text labels
}
\examples{
map_naaccr_sex_code(c(1, 2, 9))
}
