% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_path_ca.R
\name{map_path_ca}
\alias{map_path_ca}
\title{Map path_ca codes to text labels}
\usage{
map_path_ca(x)
}
\arguments{
\item{x}{A numeric (or character) vector of \code{path_ca} codes.}
}
\value{
A character vector of labels.
}
\description{
Map path_ca codes to text labels
}
\examples{
map_path_ca(c(0, 1, 10))
}
