% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_numbers_to_meaning.R
\name{redcap_numbers_to_meaning}
\alias{redcap_numbers_to_meaning}
\title{Redcap magic numbers to the data dictionary meaning.}
\usage{
redcap_numbers_to_meaning(vec, key_val_pairs, remove_names = F)
}
\arguments{
\item{vec}{The vector to operate on.}

\item{key_val_pairs}{Pre-processed key value pairs stored as a named vector.}

\item{remove_names}{Remove the names from the output vector?}
}
\value{
A vector with the conversions completed.
}
\description{
Redcap magic numbers to the data dictionary meaning.
}
