% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_splitter.R
\name{redcap_splitter}
\alias{redcap_splitter}
\title{Split PRISSMM redcap bulk download into instruments.}
\usage{
redcap_splitter(
  redcap_data,
  dict,
  keys_in_all_instr = NULL,
  forms_missing_in_redcap = NULL
)
}
\arguments{
\item{redcap_data}{The redcap bulk download (all instruments)}

\item{dict}{The data dictionary for the redcap project.}

\item{keys_in_all_instr}{Keys to keep for each split instrument.  Default is "record_id", "redcap_repeat_instrument" and "redcap_repeat_instance".}

\item{forms_missing_in_redcap}{Forms that exist in the data dictionary but come in as NA for the redcap_repeat_instrument.  Reason unknown.  Default includes "curation_initiation_eligibility", "patient_characteristics", "curation_completion", "quality_assurance".}
}
\value{
A tibble with list columns for the data dictionary subset and data instruments.
}
\description{
Split PRISSMM redcap bulk download into instruments.
}
