% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_ca_dx.R
\name{derive_ca_dx}
\alias{derive_ca_dx}
\title{Derive cancer diagnosis form}
\usage{
derive_ca_dx(raw_ca_dx, dat_dict_sub)
}
\arguments{
\item{raw_ca_dx}{Cancer diagnosis form from the raw redcap.}

\item{dat_dict_sub}{The subset of the data dictionary with variables for this instrument.}
}
\value{
A dataframe similar to ca_dx with additional derivations.
}
\description{
Derive cancer diagnosis form
}
\examples{
# Going to need synthetic data probably...
}
